+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- Added `_render_traceback_` function to improve traceback formatting of `pint.DimensionalityError`
 
 ### Changed